Split the unit test into function test and performance test
test: passed the unit test
